requirements.txt
absl-py==0.7.1
apex==0.1
apturl==0.5.2
astor==0.7.1
astunparse==1.6.3
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.4
blinker==1.4
Brlapi==0.7.0
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
command-not-found==0.3
contextlib2==0.6.0.post1
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.17
daemonize==2.4.7
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.16
duplicity==0.8.12.0
elasticsearch==6.3.1
entrypoints==0.3
enum34==1.1.6
face-alignment==1.0.0
faiss-cpu==1.6.3
fasteners==0.14.1
Flask==1.0.3
future==0.18.2
gast==0.2.2
google-auth==1.14.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
httplib2==0.14.0
idna==2.8
imageio==2.8.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jsonpatch==1.25
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==18.0.1
kiwisolver==1.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.2.2
MarkupSafe==1.1.0
matplotlib==3.2.1
mistune==0.8.4
monotonic==1.5
nbconvert==5.6.1
nbformat==5.0.6
netifaces==0.10.4
networkx==2.4
nodejs==0.1.1
notebook==6.0.3
notify2==0.3
numpy==1.18.3
oauthlib==3.1.0
olefile==0.46
opencv-python==4.2.0.34
opt-einsum==3.2.1
optional-django==0.1.0
packaging==20.3
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.7.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pkginfo==1.5.0.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycups==1.9.73
Pygments==2.6.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.16.0
python-apt==2.0.0
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyudev==0.21.0
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.3
QtPy==1.9.0
readme-renderer==26.0
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rsa==4.0
scikit-image==0.16.2
scipy==1.4.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
setproctitle==1.1.10
simplejson==3.16.0
six==1.12.0
ssh-import-id==5.10
systemd-python==234
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorboardX==2.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
torch==1.5.0
torchfile==0.1.0
torchvision==0.6.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
twine==3.1.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.24.2
usb-creator==0.3.7
visdom==0.1.8.9
wadllib==1.3.3
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xkit==0.0.0
